+++ include/igneous/engine/Time.hpp
@@ -9,6 +9,6 @@ namespace Engine
       public:
         static inline uint64_t lastTick = 0;
         static inline uint64_t currentTick = 0;
-        static inline float deltaTime = 0;
+        static inline double deltaTime = 0;
     };
 }
